Set limit=10000 for /folders/{folderId}/members
Viya 3.5 introduces version 2 of the Folders API. Both version 1 and version 2 of this API have a limit parameter, with a default of 20 results, but it did not seem to be enforced in the API at v1. It does seem to be being enforced at v2, so now specifying a limit high enough for our classroom use. For large folder structures, this may be insufficient. In this case you could increase it, or re-write this tool to iterate through several 'pages' of results. Contributions from any user welcome on this!