rule application fixes (#174)
Update to applyviyarules.py which now sees the script extract a list of existing rules from Viya and compare (using grep) the extracted rules vs the rules being applied. If the rule already exists it is discarded, if it new it is then automatically created.

If it is a new rule but contains a condition, then it is sent to an alternative csv to be manually created in the Viya env, as currently, creation of conditional rules using the CLI encounters a bug.